requirements.txt
aiohttp==3.10.10
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==5.0.0
asyncpg==0.29.0
attrs==24.2.0
black==24.10.0
braceexpand==0.1.7
click==8.1.7
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@8686b5d1bbc1d3c862292eb436ab630d6e9c9b53
discord.py==2.6.0
flake8==7.0.0
frozenlist==1.4.1
humanize==4.9.0
idna==3.10
import-expression==2.2.1.post1
jishaku==2.6.0
line-profiler==4.1.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psutil==5.9.8
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2024.1
RapidFuzz==3.10.1
six==1.16.0
toml==0.10.2
topggpy @ git+https://github.com/null8626/python-sdk@7e97f8386d0ffff8639e19a5027a7ac7e53de45f
typing_extensions==4.12.2
waifuim==1.1.1
wavelink==3.4.1
wheel==0.43.0
yarl==1.17.1